\name{snort-internal}
\alias{snort-internal}
\alias{snort.triple_store_input}
\alias{snort.triple_store_entity_register}
\alias{snort.triple_store_create_heterogeneous_links}
\alias{snort.triple_store_create_homogeneous_links_between_two_columns}
\alias{snort.triple_store_create_all_homogeneous_self_match_links}
\alias{snort.triple_store_create_homogeneous_cross_links}
\alias{snort.build_triple_store}
\alias{snort.entities_to_pajek}
\alias{snort.triples_to_pajek}
\alias{snort.triplestore_to_pajek}
\alias{snort.simple_collapse}
\alias{augment_column_names.snort}
\alias{value_counts.snort}
\alias{fastcluster.snort}
\alias{clustertables.snort}
\alias{write_cluster_tables.snort}
\docType{package}
\title{Social Network-Analysis Of Relational Tables Internal Functions only}
\description{Internal functions for SNORT}
\details{The internal functions are used to modularize the codebase and to process the data.

Internal functions:
\itemize{
\item snort.triple_store_input
\item snort.triple_store_entity_register
\item snort.triple_store_create_heterogeneous_links
\item snort.triple_store_create_homogeneous_links_between_two_columns
\item snort.triple_store_create_all_homogeneous_self_match_links
\item snort.triple_store_create_homogeneous_cross_links
\item snort.build_triple_store
\item snort.entities_to_pajek
\item snort.triples_to_pajek
\item snort.triplestore_to_pajek
\item snort.simple_collapse}

Internal Utility functions:
\itemize{
\item augment_column_names.snort
\item value_counts.snort
\item fastcluster.snort
\item clustertables.snort
\item write_cluster_tables.snort}

For details please check the source code.

N.B. These functions are internal ones only and not for general use.}
\author{Eugene Dubossarsky and Mark Norrie}
\keyword{Internal Functions}
