\name{snort}
\alias{snort}
\docType{package}
\title{Social Network-Analysis On Relational Tables}
\description{The package creates a triple store in the popular Pajek format.  It also supports a 3-Dimensional interactive VRML visualisation for network exploration.}
\details{The snort package has three main functions:
\enumerate{
\item snort.generate_all (generates all triple data in Pajek format for a fully-specified (uncollapsed) network)
\item snort.generate_collapsed (generates all triple data in Pajek format for a simple collapsed network) and
\item snort.3dgraphics (generates a 3-Dimensional interactive VRML model with fully specifiable colours and labels present or absent)}
}
\author{Eugene Dubossarsky and Mark Norrie}

